pre-0.1/requirements.txt
bokeh
configargparse
git+git://github.com/funkey/waterz
git+git://github.com/funkelab/funlib.learn.torch
git+git://github.com/funkelab/funlib.evaluate@e53a8c5099f696b076f30ef8851ebaf4c7e66ce0#egg=funlib.evaluate
git+git://github.com/funkelab/funlib.run
git+git://github.com/turagalab/malis
git+git://github.com/funkey/gunpowder@use-funlib-geometry
git+git://github.com/funkelab/lsd
git+git://github.com/funkelab/funlib.segment
git+git://github.com/funkelab/funlib.geometry
git+git://github.com/funkelab/daisy@refactor-geometry
git+git://github.com/funkelab/funlib.math
mahotas
pymongo
scipy
torch
tqdm
zarr
click
attr
cattrs
intern # for bossdb